refactor: Move pagination feature into its own class
I think I'll end up dropping this directly back into Gloubs Portal
Framework, so it's nice to have the functionality isolated.
